This PR releases Langflow 1.0 in the main branch and updates the
documentation.

+++ tests/conftest.py
@@ -1,4 +1,6 @@
 import json
+import os.path
+import shutil
 
 # we need to import tmpdir
 import tempfile
@@ -8,15 +10,19 @@ from typing import TYPE_CHECKING, AsyncGenerator
 
 import orjson
 import pytest
+from dotenv import load_dotenv
 from fastapi.testclient import TestClient
 from httpx import AsyncClient
-from sqlmodel import Session, SQLModel, create_engine
+from sqlmodel import Session, SQLModel, create_engine, select
 from sqlmodel.pool import StaticPool
 from typer.testing import CliRunner
 
 from langflow.graph.graph.base import Graph
+from langflow.initial_setup.setup import STARTER_FOLDER_NAME
 from langflow.services.auth.utils import get_password_hash
+from langflow.services.database.models.api_key.model import ApiKey
 from langflow.services.database.models.flow.model import Flow, FlowCreate
+from langflow.services.database.models.folder.model import Folder
 from langflow.services.database.models.user.model import User, UserCreate
 from langflow.services.database.utils import session_getter
 from langflow.services.deps import get_db_service
@@ -25,22 +31,53 @@ if TYPE_CHECKING:
     from langflow.services.database.service import DatabaseService
 
 
-def pytest_configure():
-    pytest.BASIC_EXAMPLE_PATH = Path(__file__).parent.absolute() / "data" / "basic_example.json"
-    pytest.COMPLEX_EXAMPLE_PATH = Path(__file__).parent.absolute() / "data" / "complex_example.json"
-    pytest.OPENAPI_EXAMPLE_PATH = Path(__file__).parent.absolute() / "data" / "Openapi.json"
-    pytest.GROUPED_CHAT_EXAMPLE_PATH = Path(__file__).parent.absolute() / "data" / "grouped_chat.json"
-    pytest.ONE_GROUPED_CHAT_EXAMPLE_PATH = Path(__file__).parent.absolute() / "data" / "one_group_chat.json"
-    pytest.VECTOR_STORE_GROUPED_EXAMPLE_PATH = Path(__file__).parent.absolute() / "data" / "vector_store_grouped.json"
+load_dotenv()
 
-    pytest.BASIC_CHAT_WITH_PROMPT_AND_HISTORY = (
-        Path(__file__).parent.absolute() / "data" / "BasicChatwithPromptandHistory.json"
-    )
-    pytest.VECTOR_STORE_PATH = Path(__file__).parent.absolute() / "data" / "Vector_store.json"
+
+def pytest_configure(config):
+    config.addinivalue_line("markers", "noclient: don't create a client for this test")
+    config.addinivalue_line("markers", "load_flows: load the flows for this test")
+    config.addinivalue_line("markers", "api_key_required: run only if the api key is set in the environment variables")
+    data_path = Path(__file__).parent.absolute() / "data"
+
+    pytest.BASIC_EXAMPLE_PATH = data_path / "basic_example.json"
+    pytest.COMPLEX_EXAMPLE_PATH = data_path / "complex_example.json"
+    pytest.OPENAPI_EXAMPLE_PATH = data_path / "Openapi.json"
+    pytest.GROUPED_CHAT_EXAMPLE_PATH = data_path / "grouped_chat.json"
+    pytest.ONE_GROUPED_CHAT_EXAMPLE_PATH = data_path / "one_group_chat.json"
+    pytest.VECTOR_STORE_GROUPED_EXAMPLE_PATH = data_path / "vector_store_grouped.json"
+    pytest.WEBHOOK_TEST = data_path / "WebhookTest.json"
+
+    pytest.BASIC_CHAT_WITH_PROMPT_AND_HISTORY = data_path / "BasicChatwithPromptandHistory.json"
+    pytest.CHAT_INPUT = data_path / "ChatInputTest.json"
+    pytest.TWO_OUTPUTS = data_path / "TwoOutputsTest.json"
+    pytest.VECTOR_STORE_PATH = data_path / "Vector_store.json"
     pytest.CODE_WITH_SYNTAX_ERROR = """
 def get_text():
     retun "Hello World"
     """
+
+    # validate that all the paths are correct and the files exist
+    for path in [
+        pytest.BASIC_EXAMPLE_PATH,
+        pytest.COMPLEX_EXAMPLE_PATH,
+        pytest.OPENAPI_EXAMPLE_PATH,
+        pytest.GROUPED_CHAT_EXAMPLE_PATH,
+        pytest.ONE_GROUPED_CHAT_EXAMPLE_PATH,
+        pytest.VECTOR_STORE_GROUPED_EXAMPLE_PATH,
+        pytest.BASIC_CHAT_WITH_PROMPT_AND_HISTORY,
+        pytest.CHAT_INPUT,
+        pytest.TWO_OUTPUTS,
+        pytest.VECTOR_STORE_PATH,
+    ]:
+        assert path.exists(), f"File {path} does not exist. Available files: {list(data_path.iterdir())}"
+
+
+@pytest.fixture(autouse=True)
+def check_openai_api_key_in_environment_variables():
+    import os
+
+    assert os.environ.get("OPENAI_API_KEY") is not None, "OPENAI_API_KEY is not set in environment variables"
 
 
 @pytest.fixture()
@@ -63,6 +100,12 @@ def session_fixture():
 class Config:
     broker_url = "redis://localhost:6379/0"
     result_backend = "redis://localhost:6379/0"
+
+
+@pytest.fixture(name="load_flows_dir")
+def load_flows_dir():
+    tempdir = tempfile.TemporaryDirectory()
+    yield tempdir.name
 
 
 @pytest.fixture(name="distributed_env")
@@ -181,26 +224,41 @@ def json_vector_store():
         return f.read()
 
 
+@pytest.fixture
+def json_webhook_test():
+    with open(pytest.WEBHOOK_TEST, "r") as f:
+        return f.read()
+
+
 @pytest.fixture(name="client", autouse=True)
-def client_fixture(session: Session, monkeypatch):
+def client_fixture(session: Session, monkeypatch, request, load_flows_dir):
     # Set the database url to a test database
-    db_dir = tempfile.mkdtemp()
-    db_path = Path(db_dir) / "test.db"
-    monkeypatch.setenv("LANGFLOW_DATABASE_URL", f"sqlite:///{db_path}")
-    monkeypatch.setenv("LANGFLOW_AUTO_LOGIN", "false")
+    if "noclient" in request.keywords:
+        yield
+    else:
+        db_dir = tempfile.mkdtemp()
+        db_path = Path(db_dir) / "test.db"
+        monkeypatch.setenv("LANGFLOW_DATABASE_URL", f"sqlite:///{db_path}")
+        monkeypatch.setenv("LANGFLOW_AUTO_LOGIN", "false")
+        if "load_flows" in request.keywords:
+            shutil.copyfile(
+                pytest.BASIC_EXAMPLE_PATH, os.path.join(load_flows_dir, "c54f9130-f2fa-4a3e-b22a-3856d946351b.json")
+            )
+            monkeypatch.setenv("LANGFLOW_LOAD_FLOWS_PATH", load_flows_dir)
+            monkeypatch.setenv("LANGFLOW_AUTO_LOGIN", "true")
 
-    from langflow.main import create_app
+        from langflow.main import create_app
 
-    app = create_app()
+        app = create_app()
 
-    # app.dependency_overrides[get_session] = get_session_override
-    with TestClient(app) as client:
-        yield client
-    # app.dependency_overrides.clear()
-    monkeypatch.undo()
-    # clear the temp db
-    with suppress(FileNotFoundError):
-        db_path.unlink()
+        # app.dependency_overrides[get_session] = get_session_override
+        with TestClient(app) as client:
+            yield client
+        # app.dependency_overrides.clear()
+        monkeypatch.undo()
+        # clear the temp db
+        with suppress(FileNotFoundError):
+            db_path.unlink()
 
 
 # create a fixture for session_getter above
@@ -241,7 +299,7 @@ def active_user(client):
             is_superuser=False,
         )
         # check if user exists
-        if active_user := session.query(User).filter(User.username == user.username).first():
+        if active_user := session.exec(select(User).where(User.username == user.username)).first():
             return active_user
         session.add(user)
         session.commit()
@@ -264,13 +322,9 @@ def flow(client, json_flow: str, active_user):
     from langflow.services.database.models.flow.model import FlowCreate
 
     loaded_json = json.loads(json_flow)
-    flow_data = FlowCreate(
-        name="test_flow",
-        data=loaded_json.get("data"),
-        user_id=active_user.id,
-        description="description",
-    )
-    flow = Flow.model_validate(flow_data.model_dump())
+    flow_data = FlowCreate(name="test_flow", data=loaded_json.get("data"), user_id=active_user.id)
+
+    flow = Flow.model_validate(flow_data)
     with session_getter(get_db_service()) as session:
         session.add(flow)
         session.commit()
@@ -280,10 +334,46 @@ def flow(client, json_flow: str, active_user):
 
 
 @pytest.fixture
-def added_flow(client, json_flow_with_prompt_and_history, logged_in_headers):
+def json_chat_input():
+    with open(pytest.CHAT_INPUT, "r") as f:
+        return f.read()
+
+
+@pytest.fixture
+def json_two_outputs():
+    with open(pytest.TWO_OUTPUTS, "r") as f:
+        return f.read()
+
+
+@pytest.fixture
+def added_flow_with_prompt_and_history(client, json_flow_with_prompt_and_history, logged_in_headers):
     flow = orjson.loads(json_flow_with_prompt_and_history)
     data = flow["data"]
     flow = FlowCreate(name="Basic Chat", description="description", data=data)
+    response = client.post("api/v1/flows/", json=flow.model_dump(), headers=logged_in_headers)
+    assert response.status_code == 201
+    assert response.json()["name"] == flow.name
+    assert response.json()["data"] == flow.data
+    return response.json()
+
+
+@pytest.fixture
+def added_flow_chat_input(client, json_chat_input, logged_in_headers):
+    flow = orjson.loads(json_chat_input)
+    data = flow["data"]
+    flow = FlowCreate(name="Chat Input", description="description", data=data)
+    response = client.post("api/v1/flows/", json=flow.model_dump(), headers=logged_in_headers)
+    assert response.status_code == 201
+    assert response.json()["name"] == flow.name
+    assert response.json()["data"] == flow.data
+    return response.json()
+
+
+@pytest.fixture
+def added_flow_two_outputs(client, json_two_outputs, logged_in_headers):
+    flow = orjson.loads(json_two_outputs)
+    data = flow["data"]
+    flow = FlowCreate(name="Two Outputs", description="description", data=data)
     response = client.post("api/v1/flows/", json=flow.model_dump(), headers=logged_in_headers)
     assert response.status_code == 201
     assert response.json()["name"] == flow.name
@@ -304,16 +394,59 @@ def added_vector_store(client, json_vector_store, logged_in_headers):
 
 
 @pytest.fixture
-def test_component_code():
-    path = Path(__file__).parent.absolute() / "data" / "component.py"
-    # load the content as a string
-    with open(path, "r") as f:
-        return f.read()
+def added_webhook_test(client, json_webhook_test, logged_in_headers):
+    webhook_test = orjson.loads(json_webhook_test)
+    data = webhook_test["data"]
+    webhook_test = FlowCreate(
+        name="Webhook Test", description="description", data=data, endpoint_name=webhook_test["endpoint_name"]
+    )
+    response = client.post("api/v1/flows/", json=webhook_test.model_dump(), headers=logged_in_headers)
+    assert response.status_code == 201
+    assert response.json()["name"] == webhook_test.name
+    assert response.json()["data"] == webhook_test.data
+    return response.json()
 
 
 @pytest.fixture
-def test_component_with_templatefield_code():
-    path = Path(__file__).parent.absolute() / "data" / "component_with_templatefield.py"
-    # load the content as a string
-    with open(path, "r") as f:
-        return f.read()
+def created_api_key(active_user):
+    hashed = get_password_hash("random_key")
+    api_key = ApiKey(
+        name="test_api_key",
+        user_id=active_user.id,
+        api_key="random_key",
+        hashed_api_key=hashed,
+    )
+    db_manager = get_db_service()
+    with session_getter(db_manager) as session:
+        if existing_api_key := session.exec(select(ApiKey).where(ApiKey.api_key == api_key.api_key)).first():
+            return existing_api_key
+        session.add(api_key)
+        session.commit()
+        session.refresh(api_key)
+    return api_key
+
+
+@pytest.fixture(name="starter_project")
+def get_starter_project(active_user):
+    # once the client is created, we can get the starter project
+    with session_getter(get_db_service()) as session:
+        flow = session.exec(
+            select(Flow)
+            .where(Flow.folder.has(Folder.name == STARTER_FOLDER_NAME))
+            .where(Flow.name == "Basic Prompting (Hello, World)")
+        ).first()
+        if not flow:
+            raise ValueError("No starter project found")
+
+        new_flow_create = FlowCreate(
+            name=flow.name,
+            description=flow.description,
+            data=flow.data,
+            user_id=active_user.id,
+        )
+        new_flow = Flow.model_validate(new_flow_create, from_attributes=True)
+        session.add(new_flow)
+        session.commit()
+        session.refresh(new_flow)
+        new_flow_dict = new_flow.model_dump()
+    return new_flow_dict
